+++ src/main/java/com/booking/superride/controller/CustomerController.java
@@ -5,6 +5,7 @@ import com.booking.superride.entity.CustomerDetails;
 import com.booking.superride.service.CustomerService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,8 +19,8 @@ public class CustomerController {
     private final CustomerService customerService;
 
     @PostMapping("/add-customer")
-    public CustomerDetails saveCustomerDetails(@RequestBody @Valid CustomerDetailsDTO customerDetailsDTO){
-        return customerService.saveCustomerDetails(customerDetailsDTO);
+    public ResponseEntity<CustomerDetails> saveCustomerDetails(@RequestBody @Valid CustomerDetailsDTO customerDetailsDTO){
+        return ResponseEntity.ok(customerService.saveCustomerDetails(customerDetailsDTO));
     }
 
 }
