+++ src/main/java/com/booking/superride/controller/RideController.java
@@ -4,6 +4,7 @@ import com.booking.superride.domain.RideBookingRequest;
 import com.booking.superride.domain.RideDetailsResponse;
 import com.booking.superride.service.RideService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,8 +18,8 @@ public class RideController {
     private final RideService rideService;
 
     @PostMapping("/book-ride")
-    public RideDetailsResponse bookRide(@RequestBody RideBookingRequest rideBookingRequest ){
-        return rideService.bookRide(rideBookingRequest);
+    public ResponseEntity<RideDetailsResponse> bookRide(@RequestBody RideBookingRequest rideBookingRequest ){
+        return ResponseEntity.ok(rideService.bookRide(rideBookingRequest));
     }
 
 }
