+++ src/main/java/com/booking/superride/domain/AddTaxiRequest.java
@@ -1,16 +1,9 @@
 package com.booking.superride.domain;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
 
-@Data
-public class AddTaxiRequest {
-
-    @NotBlank(message = "make cannot be blank")
-    private String make;
-    @NotBlank(message = "taxiNumber cannot be blank")
-    private String taxiNumber;
-    @NotBlank(message = "driverName cannot be blank")
-    private String driverName;
-
+public record AddTaxiRequest(
+        @NotBlank(message = "make cannot be blank") String make,
+        @NotBlank(message = "taxiNumber cannot be blank") String taxiNumber,
+        @NotBlank(message = "driverName cannot be blank") String driverName) {
 }
