+++ src/main/java/com/booking/superride/domain/CustomerDetailsDTO.java
@@ -1,12 +1,8 @@
 package com.booking.superride.domain;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
 
-@Data
-public class CustomerDetailsDTO {
-
-    @NotBlank(message = "customerName cannot be blank")
-    private String customerName;
-
+public record CustomerDetailsDTO(
+        @NotBlank(message = "customerName cannot be blank")
+        String customerName) {
 }
