+++ src/main/java/com/booking/superride/domain/RideBookingRequest.java
@@ -1,13 +1,10 @@
 package com.booking.superride.domain;
 
-import lombok.Data;
-
 import java.time.LocalDateTime;
 
-@Data
-public class RideBookingRequest {
-    private Long customerId;
-    private char pickupPoint;
-    private char dropPoint;
-    private LocalDateTime pickupTime;
+public record RideBookingRequest(
+        Long customerId,
+        char pickupPoint,
+        char dropPoint,
+        LocalDateTime pickupTime) {
 }
