+++ src/main/java/com/booking/superride/domain/RideDetailsResponse.java
@@ -1,14 +1,7 @@
 package com.booking.superride.domain;
 
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class RideDetailsResponse {
-
-    private String status;
-    private String message;
-    private RideDetailsDTO rideDetailsDTO;
-
+public record RideDetailsResponse(
+        String status,
+        String message,
+        RideDetailsDTO rideDetails) {
 }
