+++ src/main/java/com/booking/superride/entity/CustomerDetails.java
@@ -2,8 +2,10 @@ package com.booking.superride.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
 @Entity
 @Table(name = "customer_details")
 public class CustomerDetails {
@@ -16,4 +18,13 @@ public class CustomerDetails {
     @Column(name = "customername")
     private String customerName;
 
+    public CustomerDetails setCustomerId(Long customerId) {
+        this.customerId = customerId;
+        return this;
+    }
+
+    public CustomerDetails setCustomerName(String customerName) {
+        this.customerName = customerName;
+        return this;
+    }
 }
