+++ src/main/java/com/booking/superride/entity/TaxiDetails.java
@@ -2,8 +2,10 @@ package com.booking.superride.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
 @Entity
 @Table(name = "taxi_details")
 public class TaxiDetails {
@@ -23,4 +25,33 @@ public class TaxiDetails {
     private String driverName;
     private String status;
 
+    public TaxiDetails setTaxiId(Long taxiId) {
+        this.taxiId = taxiId;
+        return this;
+    }
+
+    public TaxiDetails setMake(String make) {
+        this.make = make;
+        return this;
+    }
+
+    public TaxiDetails setTaxiNumber(String taxiNumber) {
+        this.taxiNumber = taxiNumber;
+        return this;
+    }
+
+    public TaxiDetails setCurrentLocation(char currentLocation) {
+        this.currentLocation = currentLocation;
+        return this;
+    }
+
+    public TaxiDetails setDriverName(String driverName) {
+        this.driverName = driverName;
+        return this;
+    }
+
+    public TaxiDetails setStatus(String status) {
+        this.status = status;
+        return this;
+    }
 }
