+++ src/main/java/com/booking/superride/entity/projections/AvailableTaxiFareDetails.java
@@ -3,6 +3,6 @@ package com.booking.superride.entity.projections;
 public interface AvailableTaxiFareDetails {
 
     Long getTaxiId();
-    Double getFare();
+    double getFare();
 
 }
