+++ src/main/java/com/booking/superride/entity/projections/OnTripTaxiDetails.java
@@ -6,19 +6,4 @@ import java.time.LocalDateTime;
 public interface OnTripTaxiDetails {
 
     Long getTaxiId();
-
-    LocalDate getRideDate();
-
-    char getPickupPoint();
-
-    char getDropPoint();
-
-    long getDistance();
-
-    double getFare();
-
-    LocalDateTime getPickupTime();
-
-    LocalDateTime getDropTime();
-
 }
