--- src/main/java/com/booking/superride/entity/projections/TaxiCurrentLocationDetails.java
@@ -1,8 +1,0 @@
-package com.booking.superride.entity.projections;
-
-public interface TaxiCurrentLocationDetails {
-
-    Long getTaxiId();
-    char getCurrentLocation();
-
-}
