+++ src/main/java/com/booking/superride/repository/CustomerRepository.java
@@ -4,5 +4,4 @@ import com.booking.superride.entity.CustomerDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<CustomerDetails, Long> {
-    CustomerDetails save(CustomerDetails customerDetails);
 }
