+++ src/main/java/com/booking/superride/service/CustomerService.java
@@ -4,6 +4,7 @@ import com.booking.superride.domain.CustomerDetailsDTO;
 import com.booking.superride.entity.CustomerDetails;
 import com.booking.superride.mapper.CustomerMapper;
 import com.booking.superride.repository.CustomerRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +15,7 @@ public class CustomerService {
     private final CustomerRepository customerRepository;
     private final CustomerMapper customerMapper;
 
+    @Transactional
     public CustomerDetails saveCustomerDetails(CustomerDetailsDTO customerDetailsDTO){
         CustomerDetails details = customerMapper.dtoToDetails(customerDetailsDTO);
         return customerRepository.save(details);
