+++ src/main/java/com/booking/superride/service/TaxiService.java
@@ -2,9 +2,9 @@ package com.booking.superride.service;
 
 import com.booking.superride.domain.AddTaxiRequest;
 import com.booking.superride.domain.TaxiDetailsDTO;
-import com.booking.superride.entity.TaxiDetails;
 import com.booking.superride.mapper.TaxiMapper;
 import com.booking.superride.repository.TaxiRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,10 +15,9 @@ public class TaxiService {
     private final TaxiMapper taxiMapper;
     private final TaxiRepository taxiRepository;
 
+    @Transactional
     public TaxiDetailsDTO saveTaxiDetails(AddTaxiRequest addTaxiRequest){
-        var taxiDetails = taxiMapper.requestToDetails(addTaxiRequest);
-        taxiDetails.setCurrentLocation('A'); // Default pick up location is 'A'
-        taxiDetails.setStatus("Active"); // Default status - Active
+        var taxiDetails = taxiMapper.requestToDetails(addTaxiRequest).setCurrentLocation('A').setStatus("Active");
         taxiDetails = taxiRepository.save(taxiDetails);
         return taxiMapper.detailsToDTO(taxiDetails);
     }
