--- src/main/src/main/java/com/booking/superride/SuperRideApplication.java
@@ -1,13 +1,0 @@
-package com.booking.superride;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SuperRideApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SuperRideApplication.class, args);
-	}
-
-}
