--- src/main/src/main/java/com/booking/superride/domain/CustomerDetailsDTO.java
@@ -1,8 +1,0 @@
-package com.booking.superride.domain;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record CustomerDetailsDTO(
-        @NotBlank(message = "customerName cannot be blank")
-        String customerName) {
-}
