--- src/main/src/main/java/com/booking/superride/domain/RideBookingRequest.java
@@ -1,10 +1,0 @@
-package com.booking.superride.domain;
-
-import java.time.LocalDateTime;
-
-public record RideBookingRequest(
-        Long customerId,
-        char pickupPoint,
-        char dropPoint,
-        LocalDateTime pickupTime) {
-}
