--- src/main/src/main/java/com/booking/superride/domain/RideDetailsResponse.java
@@ -1,7 +1,0 @@
-package com.booking.superride.domain;
-
-public record RideDetailsResponse(
-        String status,
-        String message,
-        RideDetailsDTO rideDetails) {
-}
