--- src/main/src/main/java/com/booking/superride/domain/TaxiDetailsDTO.java
@@ -1,4 +1,0 @@
-package com.booking.superride.domain;
-
-public record TaxiDetailsDTO(Long taxiId, String make, String taxiNumber, char currentLocation, String driverName, String status) {
-}
