--- src/main/src/main/java/com/booking/superride/entity/projections/AvailableTaxiFareDetails.java
@@ -1,8 +1,0 @@
-package com.booking.superride.entity.projections;
-
-public interface AvailableTaxiFareDetails {
-
-    Long getTaxiId();
-    double getFare();
-
-}
