--- src/main/src/main/java/com/booking/superride/entity/projections/OnTripTaxiDetails.java
@@ -1,9 +1,0 @@
-package com.booking.superride.entity.projections;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-public interface OnTripTaxiDetails {
-
-    Long getTaxiId();
-}
