--- src/main/src/main/java/com/booking/superride/repository/CustomerRepository.java
@@ -1,7 +1,0 @@
-package com.booking.superride.repository;
-
-import com.booking.superride.entity.CustomerDetails;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CustomerRepository extends JpaRepository<CustomerDetails, Long> {
-}
