--- src/main/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.application.name=super-ride
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-spring.datasource.schema=public
